% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the amount of iterations, default is 10000}

\item{x}{the sample}

\item{fun}{the function to be applied, default is mean}

\item{alpha}{alpha, default is 0.05}

\item{cx}{for character size for text in graph, default is 1.5}

\item{...}{}
}
\value{
a graph with the point estimate with a 1 - alpha bootstrap interval
as well as a list of all the variables used
}
\description{
Bootstrap Function
}
\examples{
myboot2(x = rnorm(25,mean=25,sd=10))
}
