% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My Central Limit Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{lower limit of population}

\item{b}{upper limit of population}
}
\value{
a vector of length iter with the sample values of each iteration added together,
also creates a histogram with this vector
}
\description{
My Central Limit Function
}
\examples{
w=myclt(n=10,iter=10000
}
