% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Maximum Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to apply likelihood of}

\item{x}{vector of data}

\item{param}{vector of candidate parameter values to test}

\item{...}{any plot parameters to customize}
}
\value{
a graph of the likelihood as a function of the parameter values as well as a named list of the index of the
maximum likelihood value, the parameter value that maximizes the likelihood, the maximum likelihood value, and the slope estimate
}
\description{
My Maximum Likelihood Function
}
\examples{
mymaxlik(lfun = function(x, param) { log(dbinom(x, prob = param, size = 20)) }, x = c(3, 3, 4, 3, 4, 5, 5, 4),
param = seq(0, 1, length = 1000), xlab = expression("p"), main = "Binomial", cex.main = 2)

}
