% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the curve}

\item{sigma}{the sigma of the curve}

\item{a}{the upper bound of Y}
}
\value{
displays the curve with a shaded area between the curve and x-axis from negative infinity to x = a
         and calculate the area which is released to the command-line in a list
}
\description{
Normal Curve
}
\examples{
myncurve(10, 4, 10)

}
