% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Sample Animation}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{the amount drawn from the sample size}

\item{iter}{the number of iterations to sample and the number of graphs to produce}

\item{time}{the amount of time before a new graph is made, aka how long between each frame}
}
\value{
iter number of bar plots of each sample
}
\description{
Sample Animation
}
\examples{
mysample(n = 1000, iter = 30, time = 1)

}
