% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Selling Plane Tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats in the flight}

\item{gamma}{the probability a plane will be truly overbooked}

\item{p}{the probability of a "show"}
}
\value{
the number of tickets to be sold, displays a graph of two plots showing
the discrete and continuous distributions of the problem, as well as a named list of each variable
}
\description{
Selling Plane Tickets
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)

}
